refactor feature extractors
