+++ FE_configuration.py
@@ -4,7 +4,7 @@ from sklearn.pipeline import Pipeline
 import tsfel
 from tsfresh.feature_extraction import  EfficientFCParameters
 
-from transformers.feature_extraction import TSFELLagFeatureExtractor, TSFreshLagFeatureExtractor
+from transformers.feature_extractors import TSFELLagFeatureExtractor, TSFreshLagFeatureExtractor
 from transformers.feature_selection import FeatureSelectionWrapper
 from transformers.imputation import ImputationWrapper
 from utils import clean_column_names, restore_column_names, train_test_split_by_time
@@ -270,8 +270,8 @@ def feature_extraction_preprocessing(data, COLUMN_CONFIG, n_features, n_tsfel_fe
     
     print(subset_features_df)
     # Extract TSFEL features
-    #tsfel_config_file = extract_tsfel_features(subset_features_df, COLUMN_CONFIG, n_tsfel_features)
-    tsfel_config_file = 'tsfel_config.json' # for testing purposes
+    tsfel_config_file = extract_tsfel_features(subset_features_df, COLUMN_CONFIG, n_tsfel_features)
+    #tsfel_config_file = 'tsfel_config.json' # for testing purposes
     
     # Extract TFRESH features
     tsfresh_fc_parameters = extract_tsfresh_features(subset_features_df, COLUMN_CONFIG, n_tsfresh_features)
