+++ transformers/__init__.py
@@ -1,4 +1,4 @@
-from transformers.feature_extraction import TSFELLagFeatureExtractor, LagFeatureExtractor
+from transformers.feature_extractors import LagFeatureExtractor, TSFELLagFeatureExtractor, TSFreshLagFeatureExtractor
 from transformers.feature_selection import FeatureSelectionWrapper
 from transformers.imputation import ImputationWrapper
 from transformers.resampling import RegularityResampler
@@ -6,6 +6,7 @@ from transformers.resampling import RegularityResampler
 __all__ = [
     'LagFeatureExtractor',
     'TSFELLagFeatureExtractor',
+    'TSFreshLagFeatureExtractor',
     'FeatureSelectionWrapper',
     'ImputationWrapper',
     'RegularityResampler'
